+++ server/game/engine/BodyManager.hpp
@@ -30,7 +30,7 @@ namespace Server { namespace Game { namespace Engine {
 
         bool HasBodyType(Uint32 id) const { return id && id <= this->_bodyTypesVec.size(); }
         BodyType const& GetBodyType(Uint32 id) const { assert(this->HasBodyType(id)); return *this->_bodyTypesVec[id - 1]; }
-        Uint32 GetNbBodyTypes() const { return this->_bodyTypesVec.size(); }
+        Uint32 GetNbBodyTypes() const { return (Uint32)this->_bodyTypesVec.size(); }
 
     private:
         void _ApiRegister(Tools::Lua::CallHelper& helper);
